requirements.txt
edge-tts
ffmpeg-python
python-dotenv
rich
tqdm
yt-dlp
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
mkdocs-material
openai-whisper
pydub
nltk
pytube